requirements.txt
tensorflow==2.15.0.post1  
Pillow==10.1.0
numpy==1.24.3
keras==2.13.1
opencv-python==4.8.1.78